#18 Feat: Follower ViewModel 구현
