+++ app/src/main/java/com/sopt/now/ui/follower/FollowerFragment.kt
@@ -1,23 +1,19 @@
 package com.sopt.now.ui.follower
 
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import androidx.fragment.app.viewModels
 import com.sopt.now.databinding.FragmentFollowerBinding
-import com.sopt.now.network.response.ResponseFollowerDto
-import com.sopt.now.network.service.ServicePool
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
 
 class FollowerFragment : Fragment() {
     private var _binding: FragmentFollowerBinding? = null
     private val binding: FragmentFollowerBinding
         get() = requireNotNull(_binding) { "FragmentFollowerBinding is not initialized" }
 
+    private val viewModel: FollowerViewModel by viewModels()
     private lateinit var followerAdapter: FollowerAdapter
 
     override fun onCreateView(
@@ -32,8 +28,9 @@ class FollowerFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
+        viewModel.loadFollowers()
         setupRecyclerView()
-        loadUsers()
+        observeViewModel()
     }
 
     private fun setupRecyclerView() {
@@ -41,28 +38,10 @@ class FollowerFragment : Fragment() {
         binding.rvFollower.adapter = followerAdapter
     }
 
-    private fun loadUsers() {
-        ServicePool.followerService.getFollowers(2).enqueue(object : Callback<ResponseFollowerDto> {
-            override fun onResponse(
-                call: Call<ResponseFollowerDto>,
-                response: Response<ResponseFollowerDto>,
-            ) {
-                if (response.isSuccessful) {
-                    response.body()?.data?.let { followers ->
-                        followerAdapter.submitList(followers)
-                    }
-                } else {
-                    Log.e(
-                        "FollowerFragment",
-                        "Fail load"
-                    )
-                }
-            }
-
-            override fun onFailure(call: Call<ResponseFollowerDto>, t: Throwable) {
-                Log.e("FollowerFragment", "fetching error")
-            }
-        })
+    private fun observeViewModel() {
+        viewModel.followers.observe(viewLifecycleOwner) { followers ->
+            followerAdapter.submitList(followers)
+        }
     }
 
     override fun onDestroyView() {
